common markdown component and updated sitemap
